🚀moved search desktop template
